+++ FWApplication/Classes/include/FWApplication/FWAppConfig.h
@@ -1,1 +1,0 @@
-../../FWApplication/App/App/FWAppConfig.h